Move the SSE classes under the http package.  Add proxy classes for backwards compatibility that are marked as deprecated and will be removed in the following version.
